refactor: package rename 'utils' -> 'util'
